+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
  * `responses::TagList`, `responses::PluginList`, `responses::FeatureFlagList`, `responses::DeprecatedFeatureList`,
    `responses::MessageList` now all have `len` and `is_empty` methods.
+ * `responses::TagList`, `responses::PluginList`, `responses::NodeList`, `responses::FeatureFlagList`, `responses::DeprecatedFeatureList`,
+   `responses::MessageList` now all implement `IntoIterator`.
+
 
 ## v0.48.0 (Sep 07, 2025)